feat: refactor into separate metrics mod
