+++ uplink/src/base/serializer.rs
@@ -1,16 +1,18 @@
-use crate::{Config, Package, Point};
+use std::io;
+use std::sync::Arc;
 
 use bytes::Bytes;
 use disk::Storage;
 use flume::{Receiver, RecvError};
 use log::{error, info};
 use rumqttc::*;
-use serde::Serialize;
-use std::io;
-use std::sync::Arc;
-use std::time::{Duration, SystemTime, UNIX_EPOCH};
 use thiserror::Error;
 use tokio::{select, time};
+
+mod metrics;
+
+use crate::{Config, Package};
+use metrics::SerializerMetrics;
 
 #[derive(thiserror::Error, Debug)]
 pub enum MqttError {
@@ -162,7 +164,7 @@ pub struct Serializer<C: MqttClient> {
     collector_rx: Receiver<Box<dyn Package>>,
     client: C,
     storage: Option<Storage>,
-    metrics: Option<Metrics>,
+    metrics: Option<SerializerMetrics>,
 }
 
 impl<C: MqttClient> Serializer<C> {
@@ -183,7 +185,7 @@ impl<C: MqttClient> Serializer<C> {
             None => None,
         };
 
-        let metrics = Metrics::new(config.clone());
+        let metrics = SerializerMetrics::new(config.clone());
 
         Ok(Serializer { config, collector_rx, client, storage, metrics })
     }
@@ -501,102 +503,6 @@ async fn send_publish<C: MqttClient>(
 ) -> Result<C, MqttError> {
     client.publish_bytes(topic, QoS::AtLeastOnce, false, payload).await?;
     Ok(client)
-}
-
-#[derive(Debug, Default, Serialize, Clone)]
-pub struct Metrics {
-    #[serde(skip)]
-    topic: String,
-    sequence: u32,
-    timestamp: u64,
-    total_sent_size: usize,
-    total_disk_size: usize,
-    lost_segments: usize,
-    errors: String,
-    error_count: usize,
-}
-
-impl Metrics {
-    pub fn new(config: Arc<Config>) -> Option<Metrics> {
-        let topic = match &config.serializer_metrics.as_ref()?.topic {
-            Some(topic) => topic.to_owned(),
-            _ => {
-                String::from("/tenants/")
-                    + &config.project_id
-                    + "/devices/"
-                    + &config.device_id
-                    + "/events/"
-                    + &"metrics"
-                    + "/jsonarray"
-            }
-        };
-
-        Some(Metrics { topic, errors: String::with_capacity(1024), ..Default::default() })
-    }
-
-    pub fn add_total_sent_size(&mut self, size: usize) {
-        self.total_sent_size = self.total_sent_size.saturating_add(size);
-    }
-
-    pub fn add_total_disk_size(&mut self, size: usize) {
-        self.total_disk_size = self.total_disk_size.saturating_add(size);
-    }
-
-    pub fn sub_total_disk_size(&mut self, size: usize) {
-        self.total_disk_size = self.total_disk_size.saturating_sub(size);
-    }
-
-    pub fn increment_lost_segments(&mut self) {
-        self.lost_segments += 1;
-    }
-
-    // pub fn add_error<S: Into<String>>(&mut self, error: S) {
-    //     self.error_count += 1;
-    //     if self.errors.len() > 1024 {
-    //         return;
-    //     }
-    //
-    //     self.errors.push_str(", ");
-    //     self.errors.push_str(&error.into());
-    // }
-
-    pub fn add_errors<S: Into<String>>(&mut self, error: S, count: usize) {
-        self.error_count += count;
-        if self.errors.len() > 1024 {
-            return;
-        }
-
-        self.errors.push_str(&error.into());
-        self.errors.push_str(" | ");
-    }
-
-    pub fn update(&mut self) -> Metrics {
-        let timestamp =
-            SystemTime::now().duration_since(UNIX_EPOCH).unwrap_or(Duration::from_secs(0));
-        self.timestamp = timestamp.as_millis() as u64;
-        self.sequence += 1;
-
-        self.clone()
-    }
-
-    pub fn clear(&mut self) {
-        self.errors.clear();
-        self.lost_segments = 0;
-    }
-}
-
-impl Point for Metrics {
-    fn sequence(&self) -> u32 {
-        self.sequence
-    }
-
-    fn timestamp(&self) -> u64 {
-        self.timestamp
-    }
-
-    fn collection_timestamp(&self) -> u64 {
-        self.timestamp
-    }
 }
 
 #[cfg(test)]
